Fix Numerous Bugs
• Fixed an issue where Viridian Forest's Nuzlocke encounter slot biasing was firing incorrectly.

• Fixed vanilla tile errors on Routes 24 and 25.

• The Glass Workshop menu's exit option is now the same size as the other options.

• Fixed certain move animations playing incorrect sound effects.

• The Move Relearner should now list the correct moves in all situations.

• Displaying a Japanese Pokémon's OT name no longer crashes the game in several situations, including in the summary screen.

• Pokémon from Orre will now always be considered outsiders by the summary screen, fixing some met location printing bugs.

• Japanese Gamecube Starters and Duking's Plusle should now display their met locations correctly.

• Pokémon obtained on the S.S. Tidal should now display "FERRY" as their met location in the summary screen.

• Three Isle Path is now finished when the player gets the Extended Dex rather than the National Dex. Also fixed an issue with the prospector's text.

• Deoxys formes will now show their correct icons in the move details page of the summary screen.

• Shiny Deoxys should now display its shiny palette correctly in all situations.

+++ src/string_util.c
@@ -663,7 +663,7 @@ void ConvertInternationalString(u8 *s, u8 language)
 
         i--;
 
-        while (i != (u8)-1)
+        while (i != (u32)-1)
         {
             s[i + 2] = s[i];
             i--;
